+++ src/Plutarch/Extra/Functor.hs
@@ -1,5 +1,4 @@
 {-# LANGUAGE TypeApplications #-}
-{-# LANGUAGE TypeFamilies #-}
 
 module Plutarch.Extra.Functor (
     -- * Type classes