+++ src/Plutarch/Extra/Profunctor.hs
@@ -1,5 +1,3 @@
-{-# LANGUAGE TypeFamilies #-}
-
 module Plutarch.Extra.Profunctor (
     PProfunctor (..),
 ) where