+++ src/Plutarch/Extra/Record.hs
@@ -1,6 +1,5 @@
 {-# LANGUAGE OverloadedRecordDot #-}
 {-# LANGUAGE RankNTypes #-}
-{-# LANGUAGE TypeFamilies #-}
 
 module Plutarch.Extra.Record (
     mkRecord,